neural_pbir/requirements.txt
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git
tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
fvcore
iopath
tqdm
numpy
scipy
opencv-python
imageio
imageio-ffmpeg
scikit-image
lpips
mmcv==1.7.0
yapf==0.40.1
trimesh
PyMCubes
gpytoolbox
open3d
./neural_surface_recon/lib/cuda/
psdr-jit==0.2.1
irtk==0.1.8
largesteps
